+++ DesktopEditor/fontengine/GlyphString.h
@@ -82,7 +82,7 @@ struct TMetrics
 class TGlyph
 {
 public:
-	long         lUnicode; // Юникод
+    unsigned int lUnicode; // Юникод
 	float        fX;       // Позиция глифа
 	float        fY;       // на BaseLine
 
@@ -161,7 +161,7 @@ public:
 	~CGlyphString();
 	void SetString(const std::wstring& wsString, float fX = 0, float fY = 0);
     void SetString(const unsigned int* pGids, const unsigned int& nGidsCount, float fX = 0, float fY = 0);
-	void SetStringGID(const LONG& gid, float fX = 0, float fY = 0);
+    void SetStringGID(const unsigned int& gid, float fX = 0, float fY = 0);
 	void Reset();
 	int  GetLength();
 	void SetBBox(int nIndex, float fLeft, float fTop, float fRight, float fBottom);
